+++ src/gsIO/gsOptionList.h
@@ -15,7 +15,6 @@
 #pragma once
 
 #include <gsCore/gsForwardDeclarations.h>
-#include <gsIO/gsXml.h>
 
 namespace gismo
 {
@@ -285,30 +284,6 @@ inline std::ostream &operator<<(std::ostream &os, const gsOptionList::OptionList
 inline bool operator< ( const gsOptionList::OptionListEntry& a, const gsOptionList::OptionListEntry& b )
 { return a.label < b.label; }
 
-
-namespace internal
-{
-
-/** \brief Read OptionList from XML data
-    \ingroup IO
-*/
-template<>
-class GISMO_EXPORT gsXml<gsOptionList>
-{
-private:
-    gsXml();
-public:
-    GSXML_COMMON_FUNCTIONS(gsOptionList)
-    GSXML_GET_POINTER(gsOptionList)
-    static std::string tag () { return "OptionList"; }
-    static std::string type() { return ""; }
-
-    static void get_into(gsXmlNode * node, gsOptionList & result);
-    static gsXmlNode * put (const gsOptionList & obj, gsXmlTree & data);
-};
-
-}
-
 #ifdef GISMO_WITH_PYBIND11
 
   /**
@@ -319,3 +294,5 @@ public:
 #endif // GISMO_WITH_PYBIND11
 
 } // namespace gismo
+
+#include <gsIO/gsOptionListXml.h>
